# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FluxPay",
    "version": "0.1.0"
  },
  "paths": {
...
    "/x402/challenge": {
      "post": {
        "tags": [
          "x402 Payment Protocol"
        ],
        "summary": "Create Challenge",
        "description": "Generate RFC 7231 / x402 Payment Required response challenge.",
        "operationId": "create_challenge_x402_challenge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "402": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/x402/verify": {
      "post": {
        "tags": [
          "x402 Payment Protocol"
        ],
        "summary": "Verify Authorization",
        "description": "Verify an agent's EIP-3009 authorization signature using eth-account (no KMS needed).",
        "operationId": "verify_authorization_x402_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPayloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Verify Authorization X402 Verify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/x402/settle": {
      "post": {
        "tags": [
          "x402 Payment Protocol"
        ],
        "summary": "Settle Payment",
        "description": "Settle an authorized x402 payment and post immutable transaction to ledger.",
        "operationId": "settle_payment_x402_settle_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Settle Payment X402 Settle Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChallengeRequest": {
        "properties": {
          "resource_path": {
            "type": "string",
            "minLength": 1,
            "title": "Resource Path"
          },
          "merchant_id": {
            "type": "string",
            "minLength": 1,
            "title": "Merchant Id"
          },
          "amount_minor": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "USDC"
          }
        },
        "type": "object",
        "required": [
          "resource_path",
          "merchant_id",
          "amount_minor"
        ],
        "title": "ChallengeRequest",
        "description": "Payload to request an x402 payment challenge for a protected resource."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentPayload": {
        "properties": {
          "from": {
            "type": "string",
            "title": "From"
          },
          "to": {
            "type": "string",
            "title": "To"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Value"
          },
          "validAfter": {
            "type": "integer",
            "title": "Validafter"
          },
          "validBefore": {
            "type": "integer",
            "title": "Validbefore"
          },
          "nonce": {
            "type": "string",
            "title": "Nonce"
          },
          "v": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "V"
          },
          "r": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "R"
          },
          "s": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "S"
          },
          "signature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to",
          "value",
          "validAfter",
          "validBefore",
          "nonce"
        ],
        "title": "PaymentPayload",
        "description": "Decoded EIP-3009 transferWithAuthorization payload received in X-PAYMENT.\n\nSupports both flat authorization representations and nested authorization wrappers\nstandardized across x402 implementations."
      },
...
      "SettleRequest": {
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/PaymentPayload"
          },
          "merchant_address": {
            "type": "string",
            "title": "Merchant Address"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "payload",
          "merchant_address"
        ],
        "title": "SettleRequest",
        "description": "Payload to submit verified authorization for facilitator settlement."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyPayloadRequest": {
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/PaymentPayload"
          },
          "merchant_address": {
            "type": "string",
            "title": "Merchant Address"
          }
        },
        "type": "object",
        "required": [
          "payload",
          "merchant_address"
        ],
        "title": "VerifyPayloadRequest",
        "description": "Payload to verify an agent's EIP-3009 authorization signature."
      }
    }
  }
}